+++ openraft/src/core/raft_core.rs
@@ -44,6 +44,8 @@ use crate::error::InitializeError;
 use crate::error::QuorumNotEnough;
 use crate::error::RPCError;
 use crate::error::Timeout;
+use crate::log_id::LogIdOptionExt;
+use crate::log_id::RaftLogId;
 use crate::metrics::RaftMetrics;
 use crate::metrics::ReplicationMetrics;
 use crate::metrics::UpdateMatchedLogId;
@@ -63,8 +65,6 @@ use crate::raft::VoteRequest;
 use crate::raft::VoteResponse;
 use crate::raft::VoteTx;
 use crate::raft_state::LogStateReader;
-use crate::raft_types::LogIdOptionExt;
-use crate::raft_types::RaftLogId;
 use crate::replication::Replicate;
 use crate::replication::ReplicationCore;
 use crate::replication::ReplicationHandle;
