+++ openraft/src/lib.rs
@@ -89,6 +89,8 @@ pub use crate::defensive::DefensiveCheck;
 pub use crate::defensive::DefensiveCheckBase;
 pub use crate::entry::Entry;
 pub use crate::entry::EntryPayload;
+pub use crate::log_id::LogIdOptionExt;
+pub use crate::log_id::LogIndexOptionExt;
 pub use crate::membership::EffectiveMembership;
 pub use crate::membership::Membership;
 pub use crate::membership::StoredMembership;
@@ -104,7 +106,6 @@ pub use crate::raft::Raft;
 pub use crate::raft::RaftTypeConfig;
 pub use crate::raft_state::MembershipState;
 pub use crate::raft_state::RaftState;
-pub use crate::raft_types::LogIdOptionExt;
 pub(crate) use crate::raft_types::MetricsChangeFlags;
 pub use crate::raft_types::SnapshotId;
 pub use crate::raft_types::SnapshotSegmentId;
