+++ openraft/src/log_id/log_id_option_ext.rs
@@ -1,0 +1,33 @@
+use crate::LogId;
+use crate::NodeId;
+
+pub trait LogIdOptionExt {
+    fn index(&self) -> Option<u64>;
+    fn next_index(&self) -> u64;
+}
+
+impl<NID: NodeId> LogIdOptionExt for Option<LogId<NID>> {
+    fn index(&self) -> Option<u64> {
+        self.map(|x| x.index)
+    }
+
+    fn next_index(&self) -> u64 {
+        match self {
+            None => 0,
+            Some(log_id) => log_id.index + 1,
+        }
+    }
+}
+
+impl<NID: NodeId> LogIdOptionExt for Option<&LogId<NID>> {
+    fn index(&self) -> Option<u64> {
+        self.map(|x| x.index)
+    }
+
+    fn next_index(&self) -> u64 {
+        match self {
+            None => 0,
+            Some(log_id) => log_id.index + 1,
+        }
+    }
+}
