+++ openraft/src/raft_state/mod.rs
@@ -8,8 +8,8 @@ use crate::entry::RaftEntry;
 use crate::equal;
 use crate::error::ForwardToLeader;
 use crate::less_equal;
+use crate::log_id::RaftLogId;
 use crate::node::Node;
-use crate::raft_types::RaftLogId;
 use crate::utime::UTime;
 use crate::validate::Validate;
 use crate::LogId;
