+++ openraft/src/raft_types.rs
@@ -1,85 +1,6 @@
 use std::fmt::Display;
 use std::fmt::Formatter;
 
-use crate::vote::CommittedLeaderId;
-use crate::LogId;
-use crate::NodeId;
-
-pub trait RaftLogId<NID: NodeId> {
-    fn leader_id(&self) -> &CommittedLeaderId<NID> {
-        self.get_log_id().committed_leader_id()
-    }
-
-    fn get_log_id(&self) -> &LogId<NID>;
-
-    fn set_log_id(&mut self, log_id: &LogId<NID>);
-}
-
-pub trait LogIdOptionExt {
-    fn index(&self) -> Option<u64>;
-    fn next_index(&self) -> u64;
-}
-
-impl<NID: NodeId> LogIdOptionExt for Option<LogId<NID>> {
-    fn index(&self) -> Option<u64> {
-        self.map(|x| x.index)
-    }
-
-    fn next_index(&self) -> u64 {
-        match self {
-            None => 0,
-            Some(log_id) => log_id.index + 1,
-        }
-    }
-}
-
-impl<NID: NodeId> LogIdOptionExt for Option<&LogId<NID>> {
-    fn index(&self) -> Option<u64> {
-        self.map(|x| x.index)
-    }
-
-    fn next_index(&self) -> u64 {
-        match self {
-            None => 0,
-            Some(log_id) => log_id.index + 1,
-        }
-    }
-}
-
-pub trait LogIndexOptionExt {
-    fn next_index(&self) -> u64;
-    fn prev_index(&self) -> Self;
-    fn add(&self, v: u64) -> Self;
-}
-
-impl LogIndexOptionExt for Option<u64> {
-    fn next_index(&self) -> u64 {
-        match self {
-            None => 0,
-            Some(v) => v + 1,
-        }
-    }
-
-    fn prev_index(&self) -> Self {
-        match self {
-            None => {
-                panic!("None has no previous value");
-            }
-            Some(v) => {
-                if *v == 0 {
-                    None
-                } else {
-                    Some(*v - 1)
-                }
-            }
-        }
-    }
-
-    fn add(&self, v: u64) -> Self {
-        Some(self.next_index() + v).prev_index()
-    }
-}
-
 // Everytime a snapshot is created, it is assigned with a globally unique id.
 pub type SnapshotId = String;
