+++ openraft/src/replication/mod.rs
@@ -25,12 +25,12 @@ use crate::error::HigherVote;
 use crate::error::RPCError;
 use crate::error::ReplicationError;
 use crate::error::Timeout;
+use crate::log_id::LogIdOptionExt;
 use crate::log_id_range::LogIdRange;
 use crate::raft::AppendEntriesRequest;
 use crate::raft::AppendEntriesResponse;
 use crate::raft::InstallSnapshotRequest;
 use crate::raft::RaftMsg;
-use crate::raft_types::LogIdOptionExt;
 use crate::storage::RaftLogReader;
 use crate::storage::Snapshot;
 use crate::ErrorSubject;
